requirements.txt
flask>=3.0
xlrd>=2.0
openpyxl>=3.1
bcrypt>=4.0
python-dotenv>=1.0
gunicorn>=21.0
anthropic>=0.40
